requirements.txt
Brotli==1.0.9
click==8.0.1
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
FaBo9Axis-MPU9250==1.0.0
Flask==2.0.1
Flask-Compress==1.9.0
future==0.18.2
gpsd-py3==0.3.0
importlib-metadata==4.4.0
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.20.3
opencv-contrib-python==4.1.0.25
picamera==1.13
piicodev==0.0.3
plotly==4.14.3
psutil==5.8.0
retrying==1.3.3
RPi.GPIO==0.7.0
six==1.16.0
smbus2==0.4.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1
